Reference_old_repository.bk/deployment/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asarPy==1.0.1
asttokens==2.4.1
attrs==23.2.0
bitarray==2.9.2
black==24.2.0
cbor2==5.6.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
coverage==7.6.10
cryptography==44.0.0
cytoolz==0.12.3
dataclassy==0.11.1
eip712==0.2.4
eth-account==0.10.0
eth-brownie==1.20.6
eth-event==1.2.5
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.0.0
execnet==2.0.2
fastapi==0.115.6
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hypothesis==6.27.3
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
lru-dict==1.2.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.9.1
numpy==2.2.1
packaging==23.2
parsimonious==0.9.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
-e git+ssh://git@github.com/desibewdabbsr/pop-dev-assistant.git@a0a3be739ddc7b9da2fbc81ed9776d5ce45e17ed#egg=pop_dev_assistant
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==4.25.3
psutil==5.9.8
py==1.11.0
py-solc-ast==1.2.10
py-solc-x==1.1.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.17.2
pygments-lexer-solidity==0.7.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-forked==1.6.0
pytest-mock==3.14.0
pytest-xdist==1.34.0
python-dotenv==0.16.0
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
responses==0.25.3
rlp==4.0.0
rpds-py==0.18.0
safetensors==0.5.1
scikit-learn==1.6.0
scipy==1.15.0
semantic-version==2.10.0
setuptools==75.7.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
solc-select==1.0.4
sortedcontainers==2.4.0
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.5.1+cpu
tqdm==4.66.2
transformers==4.47.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.27.0
vvm==0.1.0
vyper==0.3.10
wcwidth==0.2.13
web3==6.15.1
websockets==12.0
wheel==0.42.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
